% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.sep.test.R
\name{auto.sep.test}
\alias{auto.sep.test}
\title{Automated separability test}
\usage{
auto.sep.test(
  ts,
  c,
  d,
  b_time,
  b_timese,
  mp_type,
  type = "not",
  ops,
  m = "MV",
  option = "each",
  fix_num = 0,
  r = 1,
  s = 1,
  per = 0,
  k = 0,
  upper = 10
)
}
\arguments{
\item{ts}{ts is the data set which is a time series data typically}

\item{c}{the maximum value of number of basis for time input}

\item{d}{the maximum value of number of basis for variate input}

\item{b_time}{type of basis for time input}

\item{b_timese}{type of basis for variate input}

\item{mp_type}{select type of mapping function, "algeb" indicates algebraic mapping on the real line. "logari" represents logarithmic mapping on the real line}

\item{type}{select type of estimation."fixt" indicates fix time t. "fixx" represents fix variate. The default is "not" which indicates no fix value}

\item{ops}{Criteria for choosing the number of bases are provided by the package, offering four options: "AIC," "BIC," "CV," and "Kfold," each corresponding to a specific Criteria}

\item{m}{the window size for the simultaneous confidence region procedure, with the default being 'MV,' which stands for the Minimum Volatility method}

\item{option}{Choose either the fixed value separability test or the for all values separability test. The default is "each"}

\item{fix_num}{fix_num indicates the use of fixed-value nonlinear time series regression. If "fixt" is chosen, it represents a fixed time value. Otherwise,
if not selected, it pertains to a fixed variate value}

\item{r}{indicates number of variate}

\item{s}{s is a positive scaling factor, the default is 1}

\item{per}{the percentage for test set used in "CV" option}

\item{k}{the number of fold used in "Kfold" option}

\item{upper}{upper The upper bound for the variate basis domain. The default value is 10. When "algeb" or "logari" is chosen, the domain is automatically set from -upper to upper}
}
\value{
If option equals to "each". A list containing dataframes with three columns each. The first column represents input values. The second column contains values of the estimated function,
along with their upper and lower bounds, which are used for separability testing. The third column is a factor indicating the types corresponding to the values
in the second column. If option equals to "all". It returns a list containing 2 lists. The first one contains SCR dataframe  and the second one is estimated function by using separating
function of t and function of x.
}
\description{
This function utilizes Simultaneous Confidence Regions (SCR) for the automated execution of separability tests with with chosen bases.
}
