% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs.plot.trans.R
\name{bs.plot.trans}
\alias{bs.plot.trans}
\title{Plots of mapping basis}
\usage{
bs.plot.trans(type, mp_type, k, upper = 10, s = 1, or = 4, title = "")
}
\arguments{
\item{type}{type indicates which type of basis is used}

\item{mp_type}{select type of mapping function, "algeb" indicates algebraic mapping on the real line. "logari" represents logarithmic mapping on the real line}

\item{k}{The k is the number of basis functions represented (If wavelet are chosen, the real number of basis is 2^k. If
Cspli is chosen, the real number of basis is k - 2 + or)}

\item{upper}{the upper bound for basis domain, the default is 10}

\item{s}{s is a positive scaling factor, the default is 1}

\item{or}{indicates the order of spline and only used in Cspli type, default is 4 which indicates cubic spline}

\item{title}{give the title for the basis plot}
}
\value{
The plot of 1 to k basis functions
}
\description{
This function generates the plot of first k basis function.
}
\examples{
bs.plot.trans("Legen", "algeb", 2)
}
