% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.plot.R
\name{cv.plot}
\alias{cv.plot}
\title{Visualization of the cross-validation results}
\usage{
cv.plot(cv_m, title = "")
}
\arguments{
\item{cv_m}{give the cross validation data frame}

\item{title}{give the title for plot}
}
\value{
the plot shows cross validation result (3D)
}
\description{
Visualization of the cross-validation results
}
