% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.res.R
\name{cv.res}
\alias{cv.res}
\title{Cross validation result by specific criteria}
\usage{
cv.res(ts, c, d, b_time, b_timese, mp_type, ops, r = 1, s = 1, per = 0, k = 0)
}
\arguments{
\item{ts}{ts is the data set which is a time series data typically}

\item{c}{the maximum value of number of basis for time input}

\item{d}{the maximum value of number of basis for variate input}

\item{b_time}{type of basis for time input}

\item{b_timese}{type of basis for variate input}

\item{mp_type}{select type of mapping function, "algeb" indicates algebraic mapping on the real line. "logari" represents logarithmic mapping on the real line}

\item{ops}{Criteria for choosing the number of bases are provided by the package, offering four options: "AIC," "BIC," "CV," and "Kfold," each corresponding to a specific Criteria}

\item{r}{indicates number of variate}

\item{s}{s is a positive scaling factor, the default is 1}

\item{per}{the percentage for test set used in "CV" option}

\item{k}{the number of fold used in "Kfold" option}
}
\value{
A data frame containing the criterion values corresponding to "c" and "d". The first element refers to the optimal number of basis for time input, and the second
element refers to the optimal number of basis for variate.
}
\description{
this function gets the cross validation result by specific criteria.
}
