% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr.plot.R
\name{scr.plot}
\alias{scr.plot}
\title{Visualization of simultaneous confidence region (SCR)}
\usage{
scr.plot(scr_df, ops, title = "", lower = -1.3, upper = 1.3)
}
\arguments{
\item{ops}{select type of estimation."nfix" refers to no fix estimation. "fixt" indicates fix time t estimation.
"fixx" represents fix variate estimation}

\item{title}{give the title for plot}

\item{lower}{give the lower bound for scale limits, the default is -1.3}

\item{upper}{give the upper bound for scale limits, the default is 1.3}

\item{res_esti}{the result of estimation}
}
\value{
the plot shows estimated function and its simultaneous confidence region (SCR)
}
\description{
Visualization of simultaneous confidence region (SCR)
}
\examples{
res_esti = fix.SCR(ts, 5, 2, m = "MV", "sin", "cos", "algeb", "fixt", 0.6, r = 2)
plot_scr(res_esti[[1]], "fixt")
}
