% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.predict.R
\name{series.predict}
\alias{series.predict}
\title{Predicting time series with 1 step}
\usage{
series.predict(
  ts,
  c,
  d,
  b_time,
  b_timese,
  mp_type,
  r = 1,
  s = 1,
  n_esti = 2000
)
}
\arguments{
\item{ts}{ts is the data set which is a time series data typically}

\item{c}{number of basis for time input}

\item{d}{number of basis for variate input}

\item{b_time}{type of basis for time input}

\item{b_timese}{type of basis for variate input}

\item{mp_type}{select type of mapping function, "algeb" indicates algebraic mapping on the real line. "logari" represents logarithmic mapping on the real line}

\item{r}{indicates number of variate}

\item{s}{s is a positive scaling factor, the default is 1}

\item{n_esti}{number of points for estimation, the default is 2000}
}
\value{
predictive values for time series
}
\description{
This function predicts the time series data basis on the estimation.
}
