% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.plot.R
\name{test.plot}
\alias{test.plot}
\title{Visulization of Simultaneous Confidence Region(SCR) for test result}
\usage{
test.plot(
  df,
  type,
  option = "each",
  ops = "",
  title = "",
  lower = -1.3,
  upper = 1.3
)
}
\arguments{
\item{df}{the result of test (estimated function under null and Simultaneous Confidence Region (SCR) )}

\item{type}{specify type of test, "homot" represents time-homogeneity test. "separa" is separability test}

\item{option}{Choose either the fixed value separability test or the for all values separability test. The default is "each"}

\item{ops}{select type of estimation."nfix" refers to no fix estimation. "fixt" indicates fix time t estimation.
"fixx" represents fix variate estimation}

\item{title}{give the title for plot}

\item{lower}{give the lower bound for scale limits, the default is -1.3}

\item{upper}{give the upper bound for scale limits, the default is 1.3}
}
\value{
the plot shows test estimated function and simultaneous confidence region (SCR)
}
\description{
Visulization of Simultaneous Confidence Region(SCR) for test result
}
